refactor: Refactor search page by separating context
