+++ apps/jurumarble/src/app/search/components/index.ts
@@ -1,5 +1,5 @@
 export { default as ChipContainer } from './ChipContainer';
-export { default as Container } from './SearchContainer';
+export { default as Container } from '../container/SearchContainer';
 export { default as DrinkInfoTopSectionItem } from './DrinkInfoTopSectionItem';
 export { default as DrinkItem } from './DrinkItem';
 export { default as DrinkInfoList } from './DrinkInfoList';
@@ -8,8 +8,9 @@ export { default as DrinkVoteItem } from './DrinkVoteItem';
 export { default as DrinkVoteList } from './DrinkVoteList';
 export { default as NoSearchResult } from './NoSearchResult';
 export { default as RegionSmallSelect } from './RegionSmallSelect';
-export { default as SearchContainer } from './SearchContainer';
+export { default as SearchContainer } from '../container/SearchContainer';
 export { default as SortSelect } from './SortSelect';
 export { default as TotalBottomSectionItem } from './TotalBottomSectionItem';
 export { default as TotalTopSectionItem } from './TotalTopSectionItem';
 export { default as VoteDescription } from './VoteDescription';
+export { default as SearchHeader } from './SearchHeader';
