+++ apps/jurumarble/src/app/search/context/index.ts
@@ -1,2 +1,3 @@
 export * from './SearchProvider';
-export * from './useSearchContext';
+export * from './useSearchValueContext';
+export * from './useSearchChangeContext';
